Changed int types to long long

+++ GeneralAlgebraTeamLabProject/Lib/task15/ri.cpp
@@ -3,28 +3,28 @@
 
 //Fq Qn Ri
 
-int NumberGCD(int a, int b) {
+long long NumberGCD(long long a, long long b) {
     if (a == 0) {
         return b;
     }
     return NumberGCD(b % a, a);
 }
 
-int m(int q, int n, int i)
+long long m(long long q, long long n, long long i)
 {
-    int m = 1;
-    while((int)pow(q,m)%(n/NumberGCD(i,n)) != 1)
+    long long m = 1;
+    while((long long)pow(q,m)%(n/NumberGCD(i,n)) != 1)
     {
         m++;
     }
     return m;
 }
 
-Polinome ri(int q, int n, int i)
+Polinome ri(long long q, long long n, long long i)
 {
-    int mValue = m(q,n,i);
+    long long mValue = m(q,n,i);
     Polinome res = Polinome();
-    for(int j = 0; j < mValue; j++)
+    for(long long j = 0; j < mValue; j++)
     {
         Polinome t = Polinome(i*pow(q,j));
         t.getCoefficients()->emplace_back(IntModulo(1));
@@ -33,22 +33,22 @@ Polinome ri(int q, int n, int i)
     return res;
 }
 
-int d(int n, int q)
+long long d(long long n, long long q)
 {
-    int d=1;
-    while((int)pow(n,d)%q != 1)
+    long long d=1;
+    while((long long)pow(n,d)%q != 1)
     {
         d++;
     }
     return d;
 }
 
-std::vector<Polinome> computePolynomialProductOfCyclotomicPlynomial(int q, int n)
+std::vector<Polinome> computePolynomialProductOfCyclotomicPlynomial(long long q, long long n)
 {
     std::vector<Polinome> res;
-    int i = 1;
+    long long i = 1;
     Polinome cyclotomic = *CyclotomicPolynomial(n,q).quotient;
-    int dValue = d(q,n);
+    long long dValue = d(q,n);
 
     bool isDivisor = false;
 
@@ -77,10 +77,10 @@ std::vector<Polinome> computePolynomialProductOfCyclotomicPlynomial(int q, int n
         }
         else
         {
-            int size = degreeOfPolinome(p);
+            long long size = degreeOfPolinome(p);
 
             Polinome r = ri(q,n,i);
-            for(int j = 0; j<q;j++)
+            for(long long j = 0; j<q;j++)
             {
                 Polinome t = p;
                 Polinome r_t = r.add(Polinome(std::to_string(j)),q);
